@@ -1,4 +1,4 @@
-import { CRCCalculator } from '../types';
+import { CRCCalculator } from '../types.js';
 
 // Generated by `./pycrc.py --algorithm=table-driven --model=jam --generate=c`
 let TABLE: Array<number> | Int32Array = [
