@@ -1,4 +1,4 @@
-import crc16 from './calculators/crc16';
-import defineCrc from './define_crc';
+import crc16 from './calculators/crc16.js';
+import defineCrc from './define_crc.js';
 
 export default defineCrc('crc-16', crc16);
