@@ -1,4 +1,4 @@
-import crc16ccitt from './calculators/crc16ccitt';
-import defineCrc from './define_crc';
+import crc16ccitt from './calculators/crc16ccitt.js';
+import defineCrc from './define_crc.js';
 
 export default defineCrc('ccitt', crc16ccitt);
