@@ -1,4 +1,4 @@
-import crcjam from './calculators/crcjam';
-import defineCrc from './define_crc';
+import crcjam from './calculators/crcjam.js';
+import defineCrc from './define_crc.js';
 
 export default defineCrc('jam', crcjam);
