@@ -1,5 +1,5 @@
-import createBuffer from './create_buffer';
-import { CRCCalculator, CRCModule } from './types';
+import createBuffer from './create_buffer.js';
+import { CRCCalculator, CRCModule } from './types.js';
 
 export default function defineCrc(model: string, calculator: CRCCalculator<Uint8Array>): CRCModule {
   const result: CRCModule = (value, previous) => calculator(createBuffer(value), previous) >>> 0;
