@@ -1,14 +1,14 @@
-import crc1 from './crc1';
-import crc8 from './crc8';
-import crc81wire from './crc81wire';
-import crc16 from './crc16';
-import crc16ccitt from './crc16ccitt';
-import crc16modbus from './crc16modbus';
-import crc16xmodem from './crc16xmodem';
-import crc16kermit from './crc16kermit';
-import crc24 from './crc24';
-import crc32 from './crc32';
-import crcjam from './crcjam';
+import crc1 from './crc1.js';
+import crc8 from './crc8.js';
+import crc81wire from './crc81wire.js';
+import crc16 from './crc16.js';
+import crc16ccitt from './crc16ccitt.js';
+import crc16modbus from './crc16modbus.js';
+import crc16xmodem from './crc16xmodem.js';
+import crc16kermit from './crc16kermit.js';
+import crc24 from './crc24.js';
+import crc32 from './crc32.js';
+import crcjam from './crcjam.js';
 
 export { crc1 };
 export { crc8 };
